Consolidate fetcher service into backend for shared database models
- Move fetcher code from fetcher/ to backend/app/fetcher/
- Enable shared database models between API and fetcher services
- Merge requirements.txt files to consolidate dependencies
- Update Docker configuration to remove separate fetcher service
- Update documentation to reflect new unified structure
- All 20 tests pass successfully

Benefits:
- Simplified project structure with single Python codebase
- Easier dependency management with unified requirements
- Better code organization under consolidated backend
- Shared database models reduce code duplication
