+++ backend/app/fetcher/__init__.py
@@ -1,0 +1,24 @@
+"""Content Fetcher Service for the News Aggregator."""
+
+from .config import settings
+from .database import (
+    create_database_tables,
+    test_database_connection,
+    get_database_session,
+    engine,
+    SessionLocal,
+    Base
+)
+from .models import Source, Article
+
+__all__ = [
+    "settings",
+    "create_database_tables", 
+    "test_database_connection",
+    "get_database_session",
+    "engine",
+    "SessionLocal",
+    "Base",
+    "Source",
+    "Article"
+]
