+++ backend/requirements.txt
@@ -5,6 +5,7 @@ alembic==1.12.1
 psycopg2-binary==2.9.9
 python-dotenv==1.0.0
 pydantic==2.5.0
+pydantic-settings==2.1.0
 httpx==0.25.2
 feedparser==6.0.10
 beautifulsoup4==4.12.2
@@ -12,3 +13,5 @@ requests==2.31.0
 apscheduler==3.10.4
 pytest==7.4.3
 pytest-asyncio==0.21.1
+asyncio-throttle==1.0.2
+python-dateutil==2.8.2
