+++ docker-compose.yml
@@ -26,18 +26,6 @@ services:
       - ./backend:/app
       - /app/__pycache__
 
-  fetcher:
-    build:
-      context: ./fetcher
-      dockerfile: ../docker/Dockerfile.fetcher
-    environment:
-      - DATABASE_URL=postgresql://newsuser:newspass@db:5432/newsdb
-    depends_on:
-      - db
-      - backend
-    volumes:
-      - ./fetcher:/app
-
   db:
     image: postgres:15-alpine
     environment:
