+++ docker/Dockerfile.fetcher
@@ -6,4 +6,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 COPY . .
 
-CMD ["python", "app/main.py"]
+CMD ["python", "-m", "app.fetcher.main"]
