--- fetcher/requirements.txt
@@ -1,14 +1,0 @@
-feedparser==6.0.10
-beautifulsoup4==4.12.2
-requests==2.31.0
-sqlalchemy==2.0.23
-psycopg2-binary==2.9.9
-python-dotenv==1.0.0
-pydantic==2.5.0
-pydantic-settings==2.1.0
-apscheduler==3.10.4
-pytest==7.4.3
-fastapi==0.104.1
-uvicorn==0.24.0
-asyncio-throttle==1.0.2
-python-dateutil==2.8.2
